fix: link lsp namespace to treesitter namespace
This commit applies the namespace styling to the Treesitter highlight, and links the LSP highlight to the Treesitter one. The benefits of this are code simplicity, not relying on an LSP for namespace styling to apply, and having a single source of truth for the highlight. The downside is that Treesitter is very generous with the namespace highlight with languages like Rust and C++, meaning more code will be italicized. However, I think this is fine since it makes some since for this code to be italicized anyway, as it is used differently in cases like this than in general cases (as a namespace to reach a different identifier).